chore(cms): move global and header footer blocks
